\name{mat_labels_fct} 
\alias{mat_labels_fct} 
\title{Builds a matrix of labels for an mvpart tree} 
\description{Builds a matrix of labels for an mvpart tree} 
 
\usage{ 
mat_labels_fct(obj)
} 
\arguments{ 
\item{obj}{an mvpart object} 
} 

\value{\item{Labels}{Matrix with pretty labels for an mvpart tree} }

\seealso{\code{\link{mvpart}}
\code{\link{MRT}}}

\keyword{Multivariate regression tree}